requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
elevenlabs
pillow
